+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Updated
 
-- Moved the Edit button the the View-Header
+- Moved the Edit & Delete buttons to the View-Header
 - Multi-Editing support for Transaction:Payee
 - Timeline chart is now also a heatmap